Narrative added / Small Fixes
Added a rough narrative to the game.

Applied a couple small fixes. Such as making it clearer on the choice selections for the player.

Made it where it would display enemy's special attack rather then just Special. Also made it where it would display the enemy's name for their HP instead of just enemy.